Merged PR 81: Undo commenting changed and Bug Fix
Somehow, the previous pull request still updated the master branch even though the pull request was rejected. This pull request is to put those changes back in. Also, fixing a bug that made the Time Expended field for Tasks not required when the Status changed to Closed.

Related work items: #372, #373