Merged PR 105: Fixing a bug in Tasks and removing Hardcoding
Fixing a bug in Tasks. When a task is closed and re-opened, then the page is refreshed or navigated away from and returned to, the Task form should now display the correct validation status for Time Expended. Also, many hardcoding pieces were replaced with more reliable code that won't break if rows are moved, removed, or added. Two more instances of hardcoding remain.

Related work items: #373, #531